Move org.apache.lucene testFixtures to io.crate.lucene namespace
To avoid split package issues with jigsaw.
Relates to https://github.com/crate/crate/issues/9819
