+++ server/src/testFixtures/java/org/elasticsearch/cluster/routing/TestShardRouting.java
@@ -19,6 +19,9 @@
 
 package org.elasticsearch.cluster.routing;
 
+import static io.crate.lucene.CrateLuceneTestCase.random;
+import static org.elasticsearch.test.ESTestCase.randomAlphaOfLength;
+
 import org.elasticsearch.Version;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.common.UUIDs;
@@ -28,9 +31,6 @@ import org.elasticsearch.repositories.IndexId;
 import org.elasticsearch.snapshots.Snapshot;
 import org.elasticsearch.snapshots.SnapshotId;
 import org.elasticsearch.test.ESTestCase;
-
-import static org.apache.lucene.tests.util.CrateLuceneTestCase.random;
-import static org.elasticsearch.test.ESTestCase.randomAlphaOfLength;
 
 /**
  * A helper that allows to create shard routing instances within tests, while not requiring to expose
