+++ server/src/testFixtures/java/org/elasticsearch/test/ESTestCase.java
@@ -59,7 +59,6 @@ import org.apache.logging.log4j.core.layout.PatternLayout;
 import org.apache.logging.log4j.status.StatusConsoleListener;
 import org.apache.logging.log4j.status.StatusData;
 import org.apache.logging.log4j.status.StatusLogger;
-import org.apache.lucene.tests.util.CrateLuceneTestCase;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.apache.lucene.tests.util.LuceneTestCase.SuppressCodecs;
 import org.apache.lucene.tests.util.TestRuleMarkFailure;
@@ -135,6 +134,7 @@ import com.carrotsearch.randomizedtesting.rules.TestRuleAdapter;
 
 import io.crate.analyze.OptimizeTableSettings;
 import io.crate.common.SuppressForbidden;
+import io.crate.lucene.CrateLuceneTestCase;
 import io.crate.testing.Asserts;
 
 /**
