+++ server/src/testFixtures/java/org/elasticsearch/test/engine/MockEngineSupport.java
@@ -31,7 +31,6 @@ import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.FilterDirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.tests.index.AssertingDirectoryReader;
-import org.apache.lucene.tests.util.CrateLuceneTestCase;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Setting.Property;
@@ -41,6 +40,8 @@ import org.elasticsearch.index.engine.EngineConfig;
 import org.elasticsearch.index.engine.EngineException;
 import org.elasticsearch.index.shard.ShardId;
 import org.elasticsearch.test.IntegTestCase;
+
+import io.crate.lucene.CrateLuceneTestCase;
 
 /**
  * Support class to build MockEngines like {@link MockInternalEngine}
